Moved admin notices to own class, reworked class organisation
